+++ fairseq/modules/sinusoidal_positional_embedding.py
@@ -59,8 +59,6 @@ class SinusoidalPositionalEmbedding(nn.Module):
         max_pos = self.padding_idx + 1 + seq_len
         if self.weights is None or max_pos > self.weights.size(0):
             # recompute/expand embeddings if needed
-            new_max_pos = max(max_pos, self.weights.size(0) if self.weights is not None else 0)
-
             self.weights = SinusoidalPositionalEmbedding.get_embedding(
                 max_pos,
                 self.embedding_dim,
